mcp public apis restartured
